Rename instance -> worker
Currently, we use "instance" in an ambiguous way, to refer either to a
load-balancing service (which may run on multiple hosts) as a whole or
individually as running on each of the hosts. This is causing confusion.

In this PR, the "individually as running on each of the hosts" meaning
is renamed to "worker".

Alternatives considered were "replica", "clone" or "node", or keeping
the current "instance" naming; on the balance, "worker" is probably the
best of these. In particular:
- "replica" suggests data replication, which is not applicable to the
  nearly-stateless HAProxy, although it would be OK;
- "node" is already used with a different meaning in Virgil and Dante;
- "clone" would be OK;
- "instance" is already used with a different meaning in the old HAProxy
  deployment and to some extent also in Iris, where the naming was
  carried over.