Avoid deploying VRRP if any of the instances have failed to deploy
If any of the individual instances failed to deploy during a rebuild, do not
deploy VRRP (keepalived) on that machine so the other machines in the clusters
can continue to serve requests.

This does not affect any pre-existing VRRP deployment; if an individual
instance fails to deploy a change, the already-running VRRP (keepalived) will
continue to run.