Use templates rather than a job matrix
This is preparatory refactoring; we want to use environments for exclusive
locking, as in Mercury, but those are only available for deployment jobs, which
do not support the matrix strategy. Therefore we switch from matrix strategy to
templates.

This is refactoring; there are some cosmetic differences, but no change to the
actual functionality expected.