Remove X-Powered-By and Server headers
These headers are optional and leak information about our webservers that could
be used by attackers. This is part of CS01283391 - WISGLOSYD - PEN. eHub
Pentest – Low/Info Severity Findings. The presence of these headers has been
flagged by security as an issue and it might be good to intercept and delete
them in HAProxy in addition to disabling them on webservers.

See also:
- haproxy !12741