Override stats ports for simulation (in addition to frontend ports)
After adding a new frontend (au2prodfrontend1sqltcp9), the simulations for unrelated frontends started failing with:

    docker: Error response from daemon: driver failed programming external connectivity on endpoint iris_au2prodfrontend1prodservices_au2lbl00345: Bind for 0.0.0.0:2031 failed: port is already allocated.

Where does this port 2031 come from, and why is it binding twice?
Well, if we look at the command (from the same error), we see that it's
binding ports 2000, 2001, and 2031:

    docker run -d --name iris_au2prodfrontend1prodservices_au2lbl00345 --user=1001:1001 --security-opt apparmor=iris_au2prodfrontend1prodservices_au2lbl00345 --restart=always -v /mnt/data1/containers/iris_au2prodfrontend1prodservices_au2lbl00345/etc:/etc/haproxy:z -v /mnt/data1/containers/iris_au2prodfrontend1prodservices_au2lbl00345/run:/var/run/haproxy:z -p 2000:2000 -p 2001:2001 -p 2031:2031 --cpus=0.5 --memory=400m wtgopstools.azurecr.io/dante/haproxy:2.4.19-bullseye -f /etc/haproxy/au2prodfrontend1prodservices.cfg

Port 2031 made sense, because the prodservices instance listed that as
its port, but it also listed some other ports:

      frontend:
        ports:
          plaintext: [2009]
          ssl: [2010]
      stats:
        ports:
          web: 2031
          socket: 2030

Why ports 2000 and 2001, then? Turns out simulation_ports is
automatically defined starting from 2000, which re-allocates the ports.
I don't know the background, but it follows the formula:

          plaintext: "{{ [(2000 + i * 2)] }}"
          ssl: "{{ [(2000 + i * 2 + 1)] }}"

This means it consumes 2 ports for each instance+machine combination
during the simulation.

However, what's relevant (I think) is that the stats ports are *not*
overridden in the simulation.

So the frontend ports grow from 2000, while the stats ports are static.
While I don't know the full inputs/outputs, it seems reasonable to
assume these two will collide at some point.

We were at 15 instance*machines, * 2 ports = 2000-2029 (30 ports).
By adding sqlreadonly, we added an extra 3 instance*machines, which
would add 6 ports = 2000-2035, which collides.

Since the ports are evidently fungible and blatantly rewritten, I feel
like just including the stats ports in the rewritten ports is reasonable
for the purposes of simulations, which is the change I've proposed here.