WI00890613: haproxy.cfg.j2: add all IP:port fields
Ensures the following fields are present:

      client_ip
      client_port
    + frontend_ip
    + frontend_port
    + backend_source_ip
      backend_source_port
    + server_ip
    + server_port

"+" means field added.

Any HAProxy request uses pair of connections, each
of which has a pair of IP:ports:

    in:
      src: client_ip:client_port
      dst: frontend_ip:frontend_port
    out:
      src: backend_source_ip:backend_source_port
      dst: server_ip:server_port

My motivation for this change originates in
billing for SQL Readonly, where we need to run the
old/new billing processors in parallel, and it's
easier to check integrity with common keys, where
they use the equivalent of [client_ip, frontend_ip].

However, this is also relevant for diagnostics,
where performance or availability investigations
might involve tracing a request path. Having a
single record with four ip:port pairs in common
would be *super* useful.

At some point during CS01802173, it was needed to
correlate a SQL query with a HAProxy request, and
we ended up guessing based on timestamps, but with
these fields it would have been easier.