WI00890613: haproxy.cfg.j2: add bytes_uploaded counter to logs
HAProxy uses %B (bytes_read: server to client) and %U (bytes_uploaded, client to server) to represent data transferred in each direction.

To get a meaningful indication of data transfer, we need to log both.

Currently, the HAProxy instances have bytes_uploaded, but Iris is missing it. This adds the field.